Move functions out of util.py
